SECURITY: Implement P0 critical cryptographic fixes
🔴 P0 Critical Fixes:

1. 64-bit Session IDs (CVE-2026-XXXXX)
   - Expanded from 32-bit to 64-bit
   - Uses full 8 bytes of HMAC-SHA256
   - Prevents birthday attack (was vulnerable at ~65k sessions)
   - Updated SessionManager, BpfSync, RateLimiter

2. Nonce Counter Manager (nonce.rs)
   - Counter-based nonce generation
   - Random base XOR atomic counter
   - Prevents nonce reuse catastrophe
t  i m-e sRteapmlpa-cbeasse d baprprookaecnh

3. Replay Protection
   - Added sequence numbers to PacketHeader
   - SessionEntry tracks last_seq
   - Framework ready for replay validation

4. Constant-Time Comparisons (crypto_util.rs)
   e- c  u 64
u    -6 4Ses
siuo nEntry:  l as t_se-q f6ie ld 4adSdeed
 s
-siuo nEntry:  l as t_se-q f6ie ld 4adSdeed
 s
-siuo nEntry:  l as t_se-q f6ie ld 4adSdeed
 s
-siuo nEntry:  l as t_se-q f6ie ld 4adSdeed
 s
-siuo nEntry: FdbEntry: session_id u64

Files Modified:
- omni-daemon/src/session.rs: 64-bit IDs, better HMAC
- omni-daemon/src/ratelimit.rs: u64 session tracking
- omni-daemon/src/bpf_sync.rs: u64 map keys
- omni-daemon/src/nonce.rs: NEW counter manager
- omni-daemon/src/crypto_util.rs: NEW constant-time ops
- omni-common/src/lib.rs: Updated wire protocol

Note: Edition 2024 reverted to 2021 (nightly required)

Status: P0 fixes complete, P1 (full Poly1305) next

+++ omni-daemon/src/ratelimit.rs
@@ -32,7 +32,7 @@ struct IpRateState {
 pub struct RateLimiter {
     config: RateLimitConfig,
     ip_states: HashMap<IpAddr, IpRateState>,
-    session_creation_times: HashMap<u32, Instant>,
+    session_creation_times: HashMap<u64, Instant>,  // Changed from u32 to u64
 }
 
 impl RateLimiter {
@@ -70,12 +70,12 @@ impl RateLimiter {
     }
 
     /// Record when a session was created (for timeout tracking).
-    pub fn record_session_start(&mut self, session_id: u32) {
+    pub fn record_session_start(&mut self, session_id: u64) {
         self.session_creation_times.insert(session_id, Instant::now());
     }
 
     /// Check if a handshake has timed out.
-    pub fn is_handshake_timeout(&self, session_id: u32) -> bool {
+    pub fn is_handshake_timeout(&self, session_id: u64) -> bool {
         if let Some(start_time) = self.session_creation_times.get(&session_id) {
             Instant::now().duration_since(*start_time) > self.config.handshake_timeout
         } else {
@@ -84,7 +84,7 @@ impl RateLimiter {
     }
 
     /// Get list of expired sessions.
-    pub fn expired_sessions(&self) -> Vec<u32> {
+    pub fn expired_sessions(&self) -> Vec<u64> {
         let now = Instant::now();
         self.session_creation_times
             .iter()
@@ -94,7 +94,7 @@ impl RateLimiter {
     }
 
     /// Remove a session from tracking.
-    pub fn remove_session(&mut self, session_id: u32) {
+    pub fn remove_session(&mut self, session_id: u64) {
         self.session_creation_times.remove(&session_id);
     }
